+++ index.html
@@ -26,7 +26,6 @@
 
     <!-- Use the api to call for a thing -->
     <nav id="mytext"></nav>
-<script src="button.js"></script>
-<script src="api_access.js"></script>
+    <script src="app.js"></script>
 </body>
 </html>
